Implement Tileset & collision fix.
I fixed collisions such that you bounce off at the exact angle that you enter and it uses the wall tilemap instead of the other object davis made.

I would like to implement the collisions in a way such that you bounce AWAY from the wall, instead of bouncing how we do currently. Like, get the normal vector of the wall and apply the car's current velocity in that direction.

I'd also like to change the car controls such that your velocity isn't tied to the way the car is facing, that way you can drift. So, for example, if you turn really fast your car has to slow down on the side you turned on before you can start moving. Conserving momentum is what I'd like to see.

I'll try working on those 2 things next before setting up the AI as it'll be based of the way the player is setup.