Close to getting the AI tracing path without the
Tried to comment explanations of what I'm doing, but basically I'm making the the enemy car trace the premade path using the mp_potential_step where it paths to a specific point and avoids objects you give it.

That's why it looks locked to the path, but the moment I give it steering and make it path using that it'll look more like a boid/car.