Made enemy car stop accelerating on turns.
I made it so that the enemy car will now stop accelerating when turning. It's pretty close to being an actual AI, the only thing I need to do next is to make sure I implement turning instead of our current mp_potential_step.

So, I'll make it turn towards the desired position using it's current speed and only accelerate when necessary. We'll see how that goes.