Got the rotation working
The enemy car will now properly turn to face the desired direction and if it overshoots it will try and make it back onto the path. It does a really good job but right now it's slowing down when it doesn't need to, and I'm doing my loop way too often and it's far too large.

I need to make it so it only slows down properly on turns, and I'll have to work on this more. But line 46 in obj_enemy/Step_0 is where you were!

So LD go back there when you get a chance.