Implement Checkpoint popup and remove "Start Race" messagebox popup
This is good to go for the playable that's due today, I won't touch it anymore for now as I need to study for 2 tests tomorrow and do some HW.

TODO:
LD needs to implement some better art for the walls so it's easier to tell that they're walls. James suggested doing a gradient which could work pretty well.

We could also just add lights along the border using our circle objects that are used to draw underglow normally so it highlights the border.

It'd also be cool if we had some sort of animation for glowing (or rather pulsing) for the walls where they'd fade to a brighter color, then fade dark, and fade back bright then loop. Then it'd look more like energy.

I also still want to do the parallax backgrounds and stuff for the tracks because it would look cool.