Implement braking
I implemented the space bar being the break button and I also made it so that when you press down to reverse that the car has to completely stop moving before you start reversing, but only if you aren't already reversing. I do the same for driving forwards.

TODO:
Fix how traction works because I drastically changed it in this push and it still isn't quite what we want.