Implement racing placements.
I implemented actual placements for the cars when racing, so you can see what place you're at and how far along the track you are.

This will make making the minimap much easier.

TODO:
Implement collision avoidance for the enemy AI so that they won't get stuck at the end of the race. They will continuously try to go to a point but if that point is occupied they should just go around and got to the nearest point instead.