Implement better collision handling for players & enemies.
TODO:
LD needs to work more on the game itself and stop focusing so much on the small issues he finds in random collision code, that's how he ends up using a whole day on just programming collisions.

TODO: Implement the FX you need for the gameplay, finish and polish the actual game loop of races, display the placements when the race is over and figure out how many racers there will be total, I'm thinking 3 enemy racers and you. That sounds good to me.

I need to polish and add lots of art to my track.