Implement FX & race finish code
Implemented drifting FX and I finished up the code for actual races. Going to work on some more so I can hand it off to James but I need to do more art finishing before then. I'd like to be able to polish my track, but if I can't it isn't a big deal.

I'd rather just make the other alternate tilesets quick and then call it a night.