Implemented optimized collision algorithm for all cars
I implemented the fully optimized collision algorithm, the game runs at 60 fps and everything now. I think I'm going to take some time to convert all the debug code to a button so that it isn't always visualized.

Then, I'm going to implement the other colors of the tilemap, the ramp, and then clean up the visuals for my track and maybe work on music. But I'm hoping james will have enough to work with so he can work on his track and I can clean his track up before the game is due.