Ramp fixes
Made it so that you "fall" to your death on the ramp if you don't have enough speed.

TODO:
Fix how checkpoints work, make it so we can handle infinite checkpoints and make it so that resetting to the last checkpoint ACTUALLY resets to the last checkpoint, instead of to a specific checkpoint.