Fixed how checkpoints work for the player so the amount of checkpoints doesn't matter.
I fixed the checkpoints so you can add as many as you like to a scene and it easily can tell if you've hit them all or not.

TODO: Fix the tutorial to include the ramp.

TODO: Implement the code for the enemy cars so they can also fall off the ramp/map.

TODO: Implement some art for the ramp.