Start implementing pausing
I need to implement the code for pausing so that I can figure out why the placement is so finnicky duing the race as the enemies cross the finish line.

TODO: Finish implementing the UI art for the race finish order and make a popup for when you cross the finish line.

Also replace the race order with the 1st, 2nd, 3rd, and 4th appearing in the top right.