Finished Implementing Pause UI
Implemented some UI for pausing so that the game is greyed out and the Pause UI shows up instead.

This works in tutorial, james_track, and circuit_city.

All the code for pausing is handled by the obj_level_manager and the obj_pause_UI objects.

the pause UI also includes the controls.

TODO:
Now that I finally can pause the game properly, I will be able to figure out why it is that when crossing the finish line the placements of cars gets messed up.