Changed how race placements are calculated
I changed how race placements are calculated in an attempt to make it so that there's no random jumps in position when cars first pass the finish line.

To continue making this more accurate I need to move the first point in the path for each track farther backwards from the finish line.