Fixed race order randomly changing
Race order would randomly change when on the checkered_obj if the track wasn't setup right, but now, we don't take the progression along the track into account when on the checkered_obj which allows us to not have sudden changes even when they shouldn't change.

This does cause a bug when the race ends where there are 2 placements for a car upon winning. this is most likely because of the way I changed the priority sorting earlier and I'm going to change it back and see if that fixes the bug while maintaining our current fixes.