Got the draw_leaderboard_GUI function fully implemented.
The leaderboard is now drawn while you are racing and when you win the race.

It uses the gamecuben font and will be drawn properly. I have to scale the time section to be a bit smaller than the rest so it fits in properly.

TODO:

I need to display the lap count when you finish each lap, and I need the lap count in the UI.

Now I'm going to make sure that the player's car is destroyed after they finish the race and we just focus on whatever car is in first instead.

Then, once all cars are finished, we can just let the player "continue" by pressing a certain button.