Implement spectating after finish.
After you finish the race, you will spectate the other racers.

TODO:
Make each track only have 4 total racers including you, include a random check on the game's start that decides which spot the player spawns at and then spawn the other racers. So, pick a random number 1-4 and that will be the slot the player spawns at, then the rest will be for the enemies.

I need to display the lap count when you finish each lap (as a popup), and I need the lap count in the UI.