Implemented lap count popup when crossing finish line.
When you cross the finish line on lap 1 it tells you that you are on lap 2, same for lap 3.

TODO:
Make each track only have 4 total racers including you, include a random check on the game's start that decides which spot the player spawns at and then spawn the other racers. So, pick a random number 1-4 and that will be the slot the player spawns at, then the rest will be for the enemies.

Add Lap count to the UI so it's always displayed.