Add the Lap count to the top middle of the GUI
TODO:
Change the way the lap times for the player is drawn in the UI so that it uses the correct font and has which lap the time is for next to it.

I also need to change it so the time defaults to 0 instead of a large number as that would work better.

Make each track only have 4 total racers including you, include a random check on the game's start that decides which spot the player spawns at and then spawn the other racers. So, pick a random number 1-4 and that will be the slot the player spawns at, then the rest will be for the enemies.