Implemented drawing the time formatted better.
Redid the way the lap times are drawn so it looks much better now.

TODO:

Make each track only have 4 total racers including you, include a random check on the game's start that decides which spot the player spawns at and then spawn the other racers. So, pick a random number 1-4 and that will be the slot the player spawns at, then the rest will be for the enemies.

At the end of a race, wait for the player to press space to go back to the main menu.