Implemented some of the game's audio.
TODO: LD forgot about the tutorial stuff that ryan redid. It'd be easier if LD Gave the game to davis right now so that he could swap that stuff out.

If he doesn't have time it's not a big deal.

Other TODO: implement the music, won't be too difficult. But, that should be the "end" of the game. LD Can go back in and spruce stuff up before we submit, cus he still has art he wants to do for both tracks, but it should be fine.